Updated love.audio.openal, and added love.audio.null. The latter will load if the OpenAL fails for some reason (if there is no audio driver on the host system, for instance). Also added __gc hooks so modules are properly destroyed. Other than that: fixed a few warnings here and there.
git-svn-id: https://love.svn.sf.net/svnroot/love/trunk@762 3494dbca-881a-0410-bd34-8ecbaf855390

+++ src/modules/graphics/opengl/Frame.cpp
@@ -42,9 +42,9 @@ namespace opengl
 		float ih = image->getHeight();
 		
 		vertices[0].x = 0; vertices[0].y = 0;
-		vertices[1].x = 0; vertices[1].y = h;
-		vertices[2].x = w; vertices[2].y = h;
-		vertices[3].x = w; vertices[3].y = 0;
+		vertices[1].x = 0; vertices[1].y = (float)h;
+		vertices[2].x = (float)w; vertices[2].y = (float)h;
+		vertices[3].x = (float)w; vertices[3].y = 0;
 		
 		vertices[0].s = x/iw; vertices[0].t = y/ih;
 		vertices[1].s = x/iw; vertices[1].t = (y+h)/ih;
